Move fixture to conftest.
