+++ tests/conftest.py
@@ -3,6 +3,7 @@ from os.path import exists, join
 
 import pytest
 
+from conda.base.context import Context
 from anaconda_anon_usage import tokens
 
 
@@ -42,3 +43,34 @@ def clear_cache():
 @pytest.fixture(autouse=True)
 def client_token_string_cache_cleanup(request):
     request.addfinalizer(clear_cache)
+
+
+@pytest.fixture(autouse=True)
+def reset_patch(request):
+    def _resetter():
+        from conda.cli import install as cli_install
+
+        for container, name, old_value in (
+            (Context, "user_agent", Context._old_user_agent),
+            (Context, "_aau_initialized", None),
+            (Context, "anaconda_anon_usage", None),
+            (
+                Context,
+                "parameter_names",
+                tuple(
+                    filter(
+                        lambda name: name != "anaconda_anon_usage",
+                        Context.parameter_names,
+                    )
+                ),
+            ),
+            (Context, "checked_prefix", None),
+            (cli_install, "check_prefix", Context._old_check_prefix),
+        ):
+            if hasattr(container, name):
+                delattr(container, name)
+            if old_value is None:
+                continue
+            setattr(container, name, old_value)
+
+    request.addfinalizer(_resetter)
